# PERF: fast inf checking in to_excel