Adds new functions to check for infinity rather than calling `np.isposinf` and `np.isneginf`, which were (surprising to me) a significant drag on `to_excel`.

I also added `xlsxwriter` to the asv build configuration.  `openpyxl` is still failing, I'm assuming something do with the specific version on conda?

```
    before     after       ratio
  [472e6e0e] [8002555d]
   156.36ms   154.16ms      0.99  packers.packers_read_excel.time_packers_read_excel
     failed     failed       n/a  packers.packers_write_excel_openpyxl.time_packers_write_excel_openpyxl
   469.69ms   357.49ms      0.76  packers.packers_write_excel_xlsxwriter.time_packers_write_excel_xlsxwriter
   368.96ms   270.60ms      0.73  packers.packers_write_excel_xlwt.time_packers_write_excel_xlwt
```