@jreback 2015-10-17
I'd rather make a bigger change to how the xls writing is done, to make it much more performant. The idea is to copy what is done for csv writing.

You chunk write, then convert the columns by blocks with ``to_native_types()`` (you would have to expand this to for example handle the ``inf`` conversions, it already handles the ``nan`` and ``float_format`` for ``FloatBlock``). these already return ``object`` arrays.

Then you simply iterate and yield the ``ExcelCells``, but no conversion is done (as it was all done before), so should be faster/simpler.