@jreback 2015-10-17
@chris-b1 agreed. Ok rebase this (prob a whatsnew conflict). The can look at other improvements if desired.