@chris-b1 2015-10-17
@jreback  - rebased